requirements.txt
Flask~=3.1.0
torch~=2.6.0
torchvision
numpy~=2.0.2
pillow~=11.1.0
ultralytics~=8.3.78
streamlit~=1.42.2
albumentations~=2.0.4
opencv-python